requirements.txt
groq
selenium
webdriver-manager
jira
requests
python-dotenv
pytest
pytest-json-report
